Feature/migration

See merge request thebrewery/flutter/globeone_poc!125

+++ lib/domain/dashboard/common/gb_converter.dart
@@ -1,5 +1,5 @@
 class GBConverter {
-  static String convert(double num) {
+  static String convert(int num) {
     const int marker = 1000;
     const int gigaBytes = marker * marker;
